developing pages
I changed the order the event from new to old.
I removed home link because it's already on the navigation bar
I changed the size of the image for aesthetic reasons.
I changed the border of EventPage's image black to white for aesthetic reason
I removed the underline of the link that goes to Events page
I replaced the positions and changed the sizes of paragraph and title of events